Extracting epochs class to its own file and reducing epoch.hpp
