+++ nano/secure/common.hpp
@@ -4,7 +4,7 @@
 #include <nano/lib/blockbuilders.hpp>
 #include <nano/lib/common.hpp>
 #include <nano/lib/config.hpp>
-#include <nano/lib/epoch.hpp>
+#include <nano/lib/epochs.hpp>
 #include <nano/lib/numbers.hpp>
 #include <nano/lib/object_stream.hpp>
 #include <nano/lib/stats.hpp>
